refactor: Add a class to encapsulate reading the files.
