+++ src/test/java/com/synopsys/integration/blackduck/developermode/DeveloperModeBdio2ReaderTest.java
@@ -1,0 +1,2 @@
+package com.synopsys.integration.blackduck.developermode;public class DeveloperModeBdio2ReaderTest {
+}
